Moving functions from convertToLTR to convertToLTRForComposer lib, renaming variable from newCommentConverted to newCommentConvertedToLTR
