+++ src/libs/convertToLTR/index.ts
@@ -3,16 +3,4 @@ import ConvertToLTR from './types';
 
 const convertToLTR: ConvertToLTR = (text) => text;
 
-const moveCursorToEndOfLine = (
-    commentLength: number,
-    setSelection: (
-        value: React.SetStateAction<{
-            start: number;
-            end: number;
-        }>,
-    ) => void,
-) => setSelection;
-
-export {moveCursorToEndOfLine};
-
 export default convertToLTR;
