+++ src/libs/convertToLTRForComposer/index.android.ts
@@ -21,4 +21,24 @@ const convertToLTRForComposer: ConvertToLTRForComposer = (text, isComposerEmpty)
     return isComposerEmpty ? `${CONST.UNICODE.LTR}${newText}` : newText;
 };
 
+/**
+ * This is necessary to convert the input to LTR, there is a delay that causes the cursor not to go to the end of the input line when pasting text or typing fast. The delay is caused for the time that takes the input to convert from RTL to LTR and viceversa.
+ */
+const moveCursorToEndOfLine = (
+    commentLength: number,
+    setSelection: (
+        value: React.SetStateAction<{
+            start: number;
+            end: number;
+        }>,
+    ) => void,
+) => {
+    setSelection({
+        start: commentLength + 1,
+        end: commentLength + 1,
+    });
+};
+
+export {moveCursorToEndOfLine};
+
 export default convertToLTRForComposer;
