+++ src/pages/home/report/ReportActionCompose/ComposerWithSuggestions/ComposerWithSuggestions.js
@@ -14,8 +14,7 @@ import canFocusInputOnScreenFocus from '@libs/canFocusInputOnScreenFocus';
 import compose from '@libs/compose';
 import * as ComposerUtils from '@libs/ComposerUtils';
 import getDraftComment from '@libs/ComposerUtils/getDraftComment';
-import {moveCursorToEndOfLine} from '@libs/convertToLTR';
-import convertToLTRForComposer from '@libs/convertToLTRForComposer';
+import convertToLTRForComposer, {moveCursorToEndOfLine} from '@libs/convertToLTRForComposer';
 import * as EmojiUtils from '@libs/EmojiUtils';
 import focusComposerWithDelay from '@libs/focusComposerWithDelay';
 import * as KeyDownListener from '@libs/KeyboardShortcut/KeyDownPressListener';
@@ -223,18 +222,18 @@ function ComposerWithSuggestions({
                 }
             }
 
-            let newCommentConverted = newComment;
+            let newCommentConvertedToLTR = newComment;
             const prevComment = commentRef.current;
 
             // This prevent the double execution of setting input value that could affect the place holder and could send an empty message or draft messages in android
             if (prevComment !== newComment) {
-                newCommentConverted = convertToLTRForComposer(newCommentConverted, composerIsEmpty.current);
-                setValue(newCommentConverted);
+                newCommentConvertedToLTR = convertToLTRForComposer(newCommentConvertedToLTR, composerIsEmpty.current);
+                setValue(newCommentConvertedToLTR);
                 moveCursorToEndOfLine(newComment.length, setSelection);
                 composerIsEmpty.current = false;
             }
 
-            const isNewCommentEmpty = !!newCommentConverted.match(/^(\s)*$/);
+            const isNewCommentEmpty = !!newCommentConvertedToLTR.match(/^(\s)*$/);
             const isPrevCommentEmpty = !!prevComment.match(/^(\s)*$/);
 
             /** Only update isCommentEmpty state if it's different from previous one */
@@ -256,22 +255,22 @@ function ComposerWithSuggestions({
             }
 
             // Indicate that draft has been created.
-            if (prevComment.length === 0 && newCommentConverted.length !== 0) {
+            if (prevComment.length === 0 && newCommentConvertedToLTR.length !== 0) {
                 Report.setReportWithDraft(reportID, true);
             }
 
             // The draft has been deleted.
-            if (newCommentConverted.length === 0) {
+            if (newCommentConvertedToLTR.length === 0) {
                 Report.setReportWithDraft(reportID, false);
             }
 
-            commentRef.current = newCommentConverted;
+            commentRef.current = newCommentConvertedToLTR;
             if (shouldDebounceSaveComment) {
-                debouncedSaveReportComment(reportID, newCommentConverted);
+                debouncedSaveReportComment(reportID, newCommentConvertedToLTR);
             } else {
-                Report.saveReportComment(reportID, newCommentConverted || '');
+                Report.saveReportComment(reportID, newCommentConvertedToLTR || '');
             }
-            if (newCommentConverted) {
+            if (newCommentConvertedToLTR) {
                 debouncedBroadcastUserIsTyping(reportID);
             }
         },
